requirements.txt
torch>=2.0.0
torchvision>=0.15.0
datasets>=2.14.0
geopy>=2.4.0
numpy>=1.24.0
Pillow>=9.5.0
huggingface_hub>=0.16.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
